backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
mangum==0.17.0
python-dotenv==1.0.0
openai>=1.40.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
# Removed: boto3==1.34.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
google-genai>=1.39.0
google-cloud-firestore>=2.16.0